record-tgenerator-api/requirements.txt
numpy
scikit-learn
matplotlib
opencv-python
tqdm
tensorflow==2.20
pillow
fastapi[standard]
uvicorn
pyyaml